@@ -1,44 +1,27 @@
 {
-    "private": true,
-    "scripts": {
-        "start": "webpack serve",
-        "build": "webpack"
-    },
-    "dependencies": {
-        "@fortawesome/fontawesome-free": "^5.13.1",
-        "exceljs": "^4.4.0",
-        "moment": "^2.29.2",
-        "react": "^17.0.2",
-        "react-dom": "^17.0.2",
-        "remotedev": "^0.2.9"
-    },
-    "devDependencies": {
-        "@babel/core": "^7.26.0",
-        "@babel/plugin-transform-runtime": "^7.25.9",
-        "@babel/polyfill": "^7.12.1",
-        "@babel/preset-env": "^7.26.0",
-        "@babel/preset-react": "^7.26.3",
-        "@babel/runtime": "^7.26.0",
-        "autoprefixer": "^10.4.20",
-        "babel-loader": "^8.4.1",
-        "copy-webpack-plugin": "^9.0.0",
-        "core-js": "^3.15.1",
-        "css-loader": "^5.2.6",
-        "file-loader": "^6.2.0",
-        "html-webpack-plugin": "^5.3.2",
-        "mini-css-extract-plugin": "^1.6.0",
-        "postcss": "^8.4.49",
-        "postcss-loader": "^8.1.1",
-        "sass": "^1.35.1",
-        "sass-loader": "^12.1.0",
-        "style-loader": "^3.0.0",
-        "tailwindcss": "^3.4.17",
-        "terser": "^5.7.0",
-        "webpack": "^5.97.1",
-        "webpack-cli": "^4.7.2",
-        "webpack-dev-server": "^3.11.2"
-    },
-    "resolutions": {
-        "loader-utils": "2.0.4"
-    }
+  "name": "musiorder-client",
+  "version": "0.0.1",
+  "private": true,
+  "scripts": {
+    "start": "dotnet tool restore && dotnet fable watch src --runFast vite",
+    "build": "dotnet tool restore && dotnet fable src --run vite build",
+    "clean": "dotnet fable clean src --yes"
+  },
+  "dependencies": {
+    "@fortawesome/fontawesome-free": "^6.7.2",
+    "@tailwindcss/vite": "^4.0.9",
+    "exceljs": "^4.4.0",
+    "moment": "^2.30.1",
+    "react": "^19.0.0",
+    "react-dom": "^19.0.0",
+    "tailwindcss": "^4.0.9",
+    "use-sync-external-store": "^1.4.0"
+  },
+  "devDependencies": {
+    "@vitejs/plugin-react": "^4.3.4",
+    "vite": "^6.2.0"
+  },
+  "engines": {
+    "node": ">=18"
+  }
 }
