@@ -2,4 +2,4 @@ module JS
 
 open Fable.Core
 
-let [<Import("*","moment")>] moment: System.DateTimeOffset -> obj = jsNative
+let [<Import("default","moment")>] moment: System.DateTimeOffset -> obj = jsNative
